+++ packages/expo/cli/start/server/__mocks__/DevelopmentSession.ts
@@ -1,0 +1,6 @@
+export class DevelopmentSession {
+  constructor(public projectRoot: string, public url: string) {}
+
+  startAsync = jest.fn(async () => ({}));
+  stop = jest.fn();
+}